pyproject.toml
[build-system]
requires = ["uv_build>=0.9.15"]
build-backend = "uv_build"

[project]
name = "xclients"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "av>=16.0.1",
    "einops>=0.8.1",
    "gradio>=6.6.0",
    "jaxtyping>=0.3.3",
    "matplotlib>=3.10.8",
    "numpy>=2.3.5",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "pydantic>=2.12.4",
    "pyrender>=0.1.45",
    "rerun-sdk>=0.27.2",
    "rich>=14.2.0",
    "scipy>=1.16.3",
    "torch>=2.10.0",
    "trimesh>=4.10.0",
    "tyro>=0.9.35",
    "webpolicy",
    "xarm-python-sdk>=1.17.3",
    "yourdfpy>=0.0.58",
]

[tool.uv.sources]
webpolicy = { git = "https://github.com/mhyatt000/webpolicy" }
xarm-python-sdk = { git = "https://github.com/xArm-Developer/xArm-Python-SDK" }


# [project.scripts]
# yourpkg = "yourpkg.cli:main"   # if you expose a CLI
# da3 = "depth_anything_3.cli:app"

[tool.pytest.ini_options]
addopts = "-q"

# [tool.ruff]
# target-version = "py310"
# line-length = 100
# lint.extend-select = ["I", "UP", "B", "C4", "SIM", "PERF"]
# fix = true
# src = ["src", "tests"]

[tool.pyright]
include = ["src", "tests"]
reportMissingTypeStubs = false

[tool.coverage.run]
source = ["yourpkg"]
branch = true
